Fix VERSION file and volcreate switches (#7)
* VERSION: update to 2.6

c8e2f42f3589d4045749deaf4eef85df30d06b28 ("Code and POD cleanup,
release 2.6 ...") updated the version number in all of the scripts
to 2.6 but failed to update the VERSION file.  This change updates
VERSION for consistency.

* volcreate: change --quiet --quota --version --volume aliases

e50265e9ec616c5f6b9fae79e8ff290b7a727aa0 ("Add support for Auristor,
reformat Perl") introduced the ability to use command line switches.
However, it assigned -q as an alias for both --quiet and --quota
and it assigned -v as an alias for --volume where -v is used as an
alias for --version in all of the other commands.

This change introduces -Q for --quota and -V for --volume; -v can
then be used for --version.